requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
backcall=0.1.0=pypi_0
blas=1.0=mkl
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py38_0
chardet=3.0.4=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cython=0.29.16=pypi_0
decorator=4.4.2=pypi_0
gluonnlp=0.9.1=pypi_0
idna=2.9=pypi_0
intel-openmp=2020.0=166
ipython=7.13.0=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.17.0=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
mkl=2020.0=166
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.0.15=py38ha843d7b_0
mkl_random=1.1.0=py38h962f231_0
mxnet=1.6.0=pypi_0
ncurses=6.2=he6710b0_0
ninja=1.9.0=py38hfd86e86_0
numpy=1.18.1=py38h4f9e942_0
numpy-base=1.18.1=py38hde5b4d6_1
openssl=1.1.1f=h7b6447c_0
packaging=20.3=pypi_0
parso=0.7.0=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pip=20.0.2=py38_1
prompt-toolkit=3.0.5=pypi_0
ptyprocess=0.6.0=pypi_0
pygments=2.6.1=pypi_0
pyparsing=2.4.7=pypi_0
python=3.8.2=hcf32534_0
python-graphviz=0.8.4=pypi_0
pytorch=1.4.0=py3.8_cuda10.1.243_cudnn7.6.3_0
readline=8.0=h7b6447c_0
requests=2.23.0=pypi_0
setuptools=46.1.3=py38_0
six=1.14.0=py38_0
sqlite=3.31.1=h7b6447c_0
tk=8.6.8=hbc83047_0
tqdm=4.45.0=pypi_0
traitlets=4.3.3=pypi_0
urllib3=1.25.9=pypi_0
wcwidth=0.1.9=pypi_0
wheel=0.34.2=py38_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
